move stuff in utils into comp_decomp
Also documents our new pack_bytes function
